added proper authorization to routes

+++ app/api/board_routes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, request
 from app.models import Board, db, Section
-from flask_login import login_required
+from flask_login import login_required, current_user
 from .auth_routes import validation_errors_to_error_messages
 from ..forms.create_board_form import CreateBoardForm
 
@@ -12,14 +12,20 @@ board_routes = Blueprint('boards', __name__, url_prefix="/api/boards")
 @login_required
 # Get all sections by board id
 def get_sections(board_id):
-    # Query all sections by board id
-    sections = Section.query.filter(Section.board_id,)
-    return { 'sections': [section.to_dict() for section in sections] }
+    # Query board by board id
+    board = Board.query.get(board_id)
+    if (board.user_id == current_user.to_dict().id):
+        # Query all sections by board id
+        sections = Section.query.filter(Section.board_id == board_id)
+        return { 'sections': [section.to_dict() for section in sections] }
+    else:
+        return {'errors': ['Unauthorized']}, 401
 
-@board_routes.route('/<int:user_id>', methods=["POST"])
+@board_routes.route('', methods=["POST"])
 @login_required
 # Create a board of current user
-def create_board(user_id):
+def create_board():
+    user_id = current_user.to_dict().id
     # Creates instance of create board form class
     form = CreateBoardForm()
     form['csrf_token'].data = request.cookies['csrf_token']
@@ -41,9 +47,13 @@ def create_board(user_id):
 @login_required
 # Get single board by id
 def get_board(board_id):
-    # Query all boards by board id
+    # Query board by board id
     board = Board.query.get(board_id)
-    return { "Board": board.to_dict() }
+    # Check if board belongs to user
+    if (board.user_id == current_user.to_dict().id):
+        return { "Board": board.to_dict() }
+    else:
+        return {'errors': ['Unauthorized']}, 401
 
 @board_routes.route('/<int:board_id>', methods=["DELETE"])
 @login_required
@@ -51,8 +61,12 @@ def get_board(board_id):
 def delete_board(board_id):
     # Query a board by board id
     board = Board.query.get(board_id)
-    # Deletes board from database
-    db.session.delete(board)
-    # Updates database
-    db.session.commit()
-    return {'message': 'Successfully deleted!'}
+    # Check if board belongs to user
+    if (board.user_id == current_user.to_dict().id):
+        # Deletes board from database
+        db.session.delete(board)
+        # Updates database
+        db.session.commit()
+        return {'message': 'Successfully deleted!'}
+    else:
+        return {'errors': ['Unauthorized']}, 401
